+++ roboRIO/src/main/java/frc/team2225/robot/Robot.java
@@ -4,6 +4,7 @@ import edu.wpi.first.wpilibj.SPI;
 import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj.command.Scheduler;
 import frc.team2225.robot.subsystem.Drivetrain;
+import frc.team2225.robot.subsystem.RollerIntake;
 
 /**
  * The VM is configured to automatically run this class, and to call the
@@ -14,6 +15,7 @@ import frc.team2225.robot.subsystem.Drivetrain;
  */
 public class Robot extends TimedRobot {
   public static Drivetrain drivetrain;
+  public static RollerIntake rollerIntake;
   /**
    * This function is run when the robot is first started up and should be
    * used for any initialization code.
@@ -21,6 +23,8 @@ public class Robot extends TimedRobot {
   @Override
   public void robotInit() {
     drivetrain = new Drivetrain(4, 5, 2, 6, SPI.Port.kOnboardCS0);
+    //TODO: Find the IDs for the roller intake motors
+    rollerIntake = new RollerIntake(0,1);
   }
 
   /**
